Move Finish component to ts
